refactor NightlyRunner to be main class
