+++ build.gradle
@@ -200,7 +200,6 @@ dependencies {
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter', version: springBoot.class.package.implementationVersion
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-aop', version: springBoot.class.package.implementationVersion
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-json', version: springBoot.class.package.implementationVersion
-  implementation group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: springBoot.class.package.implementationVersion
 
   implementation group: 'org.springframework.cloud', name: 'spring-cloud-starter-openfeign'
   implementation group: 'org.springframework.cloud', name: 'spring-cloud-starter-bootstrap'
@@ -243,7 +242,7 @@ dependencyManagement {
 }
 
 application {
-  mainClass.set('uk.gov.hmcts.reform.sscs.trigger.Application')
+  mainClass.set('uk.gov.hmcts.reform.sscs.trigger.NightlyRunner')
 }
 
 bootJar {
