--- src/main/java/uk/gov/hmcts/reform/sscs/trigger/Application.java
@@ -1,24 +1,0 @@
-package uk.gov.hmcts.reform.sscs.trigger;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.ComponentScan;
-import uk.gov.hmcts.reform.idam.client.IdamApi;
-
-@SpringBootApplication
-@ComponentScan(basePackages = {
-    "uk.gov.hmcts.reform.idam.client",
-    "uk.gov.hmcts.reform.ccd.client",
-    "uk.gov.hmcts.reform.sscs"})
-@EnableFeignClients(clients = { IdamApi.class })
-@SuppressWarnings("HideUtilityClassConstructor")
-public class Application {
-
-    public static void main(final String[] args) {
-        ConfigurableApplicationContext context = SpringApplication.run(Application.class, args);
-        int exitCode = SpringApplication.exit(context);
-        System.exit(exitCode);
-    }
-}
