+++ src/main/java/uk/gov/hmcts/reform/sscs/trigger/NightlyRunner.java
@@ -2,6 +2,10 @@ package uk.gov.hmcts.reform.sscs.trigger;
 
 import lombok.extern.log4j.Log4j2;
 import org.springframework.boot.CommandLineRunner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.ccd.client.CaseEventsApi;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
@@ -11,6 +15,7 @@ import uk.gov.hmcts.reform.ccd.client.model.CaseEventDetail;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
 import uk.gov.hmcts.reform.ccd.client.model.SearchResult;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
+import uk.gov.hmcts.reform.idam.client.IdamApi;
 import uk.gov.hmcts.reform.sscs.trigger.service.AuthorisationService;
 import uk.gov.hmcts.reform.sscs.trigger.triggers.Trigger;
 
@@ -20,6 +25,12 @@ import static java.util.Objects.requireNonNull;
 
 @Component
 @Log4j2
+@SpringBootApplication
+@ComponentScan(basePackages = {
+    "uk.gov.hmcts.reform.idam.client",
+    "uk.gov.hmcts.reform.ccd.client",
+    "uk.gov.hmcts.reform.sscs"})
+@EnableFeignClients(clients = { IdamApi.class })
 @SuppressWarnings("PMD.DoNotTerminateVM")
 public class NightlyRunner implements CommandLineRunner {
 
@@ -40,6 +51,10 @@ public class NightlyRunner implements CommandLineRunner {
         this.caseEventsApi = caseEventsApi;
         this.triggers = triggers;
         this.authorisationService = authorisationService;
+    }
+
+    public static void main(final String[] args) {
+        SpringApplication.run(NightlyRunner.class, args);
     }
 
     @Override
